% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_index_plot}
\alias{add_index_plot}
\title{Add index to plot}
\usage{
add_index_plot(.plot, method, index)
}
\arguments{
\item{.plot}{The \code{ggplot2} object derived from results.}

\item{method}{The critical power method used.}

\item{index}{The index of the combination}
}
\value{
A \code{ggplot2} object
}
\description{
Internal use only.
}
\keyword{internal}
