% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_critical_power}
\alias{demo_critical_power}
\title{Critical Power Dataset}
\format{
A data frame with 5 rows and 2 variables:
\describe{
\item{PO}{power output values, in watts.}
\item{TTE}{time-to-exhaustion, in seconds, from the corresponding power output.}
}
}
\usage{
demo_critical_power
}
\description{
A dataset containing 5 time-to-exhaustion trials for demonstrating
how to use the \code{critical_power()} function.
}
\keyword{datasets}
