% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_critical_speed}
\alias{demo_critical_speed}
\title{Critical Speed Dataset}
\format{
A data frame with 3 rows and 2 variables:
\describe{
\item{Distance}{distance values, in meters}
\item{TTE}{time-to-exhaustion, in seconds, from the corresponding power output.}
}
}
\usage{
demo_critical_speed
}
\description{
A dataset containing 5 time-to-exhaustion trials for demonstrating
how to use the \code{critical_speed()} function.
}
\keyword{datasets}
