% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_data}
\alias{demo_data}
\title{Dataset for demonstration purposes}
\format{
A data frame with 5 rows and 2 variables:
\describe{
\item{PO}{power output values, in watts.}
\item{TTE}{time-to-exhaustion, in seconds, from the corresponding power output.}
}
}
\usage{
demo_data
}
\description{
A dataset containing 5 time-to-exhaustion trials for demonstrating
how to use the package.
}
\keyword{datasets}
