% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_combinations}
\alias{get_combinations}
\title{Get combinations}
\usage{
get_combinations(.data, power_output_column, time_to_exhaustion_column)
}
\arguments{
\item{.data}{A data frame containing a power output column and their respective time-to-exhaustion.}

\item{power_output_column}{The name of the power output column. This value has to be in watts. Default to \code{"PO"}.}

\item{time_to_exhaustion_column}{The name of the time-to-exhaustion column. This value has to be in seconds. Default to \code{"TTE"}.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Derives all the possible combinations of the given time-to-exhaustion trials.
This function is mainly for internal use.
}
