% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_indexes}
\alias{get_indexes}
\title{Get index of combinations}
\usage{
get_indexes(
  .data,
  power_output_column,
  time_to_exhaustion_column,
  combinations
)
}
\arguments{
\item{.data}{A data frame containing a power output column and their respective time-to-exhaustion.}

\item{power_output_column}{The name of the power output column. This value has to be in watts. Default to \code{"PO"}.}

\item{time_to_exhaustion_column}{The name of the time-to-exhaustion column. This value has to be in seconds. Default to \code{"TTE"}.}

\item{combinations}{The data derived from \code{get_combinations()}.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
This function is used in combination with \code{get_combinations()} and it derives the indexes
of the time-to-exhaustion trials in the different combinations.
}
\examples{
combinations <- get_combinations(demo_data, "PO", "TTE")

results_index <- get_indexes(demo_data, "PO", "TTE", combinations)

results_index
}
