% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-cp.R
\name{method_1_time}
\alias{method_1_time}
\title{CP 1/time linear model}
\usage{
method_1_time(
  .data,
  power_output_column,
  time_to_exhaustion_column,
  plot = TRUE
)
}
\arguments{
\item{.data}{A data frame containing a power output column and their respective time-to-exhaustion.}

\item{power_output_column}{The name of the power output column. This value has to be in watts. Default to \code{"PO"}.}

\item{time_to_exhaustion_column}{The name of the time-to-exhaustion column. This value has to be in seconds. Default to \code{"TTE"}.}

\item{plot}{A boolean indicating whether to produce a plot from the critical power estimation. Default to \code{TRUE}.}
}
\value{
a \link[tibble:tibble-package]{tibble} containing the following columns:
\item{method}{The critical power method for that estimation.}
\item{data}{The data you provided.}
\item{model}{A \code{nls} or \code{lm} object. The model used in the critical power fitting.}
\item{CP}{Critical power, in watts.}
\item{CP SEE}{Standard error of the estimation of critical power, in watts.}
\item{W'}{Anaerobic work capacity, in joules.}
\item{W' SEE}{Standard error of the estimation of anaerobic work capacity, in joules.}
\item{Pmax}{Maximal instantaneous power, in watts (only available in CP 3-hyp).}
\item{Pmax SEE}{Standard error of the estimation of maximal instantaneous power, in watts (only available in CP 3-hyp).}
\item{R2}{R-squared.}
\item{RMSE}{Root mean square error, the units vary depending on the method, as they represent the units of the response (i.e., y-axis): CP3-hyp in seconds, CP2-hyp in seconds, CPlinear in joules, and CP1/time in watts.}
\item{plot}{The critical power plot.}
}
\description{
Fits critical power with the 1/time linear model (Whipp et al. 1982).
For more details see \code{?critical_power}.
}
\examples{
results_1_time <- method_1_time(.data = demo_data, "PO", "TTE")

results_1_time
}
