% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-cp.R
\name{method_cp}
\alias{method_cp}
\title{CP models}
\usage{
method_cp(
  .data,
  power_output_column,
  distance_column,
  time_to_exhaustion_column = "TTE",
  method = c("3-hyp", "2-hyp", "linear", "1/time"),
  mode = c("power", "speed"),
  plot = TRUE,
  reverse_y_axis = FALSE
)
}
\arguments{
\item{.data}{A data frame containing a power output/distance column and their respective time-to-exhaustion.}

\item{power_output_column}{The name of the power output column. This value has to be in watts. Used in the 'power' mode only.}

\item{distance_column}{The name of the distance column. This value has to be in meters. Used in the 'speed' mode only.}

\item{time_to_exhaustion_column}{The name of the time-to-exhaustion column. This value has to be in seconds. Default to \code{"TTE"}.}

\item{method}{The method for estimating critical speed. It can be one or multiple methods. Default to \code{c("3-hyp", "2-hyp", "linear", "1/time")}.}

\item{mode}{Specify 'power' for critical power, and 'speed' for critical speed.}

\item{plot}{A boolean indicating whether to produce a plot from the critical power/speed estimation. Default to \code{TRUE}.}

\item{reverse_y_axis}{A boolean to indicate whether to plot the Power Output (W) or Speed (m/s) in the y-axis. It is ignored for the linear methods. Default to \code{FALSE}.}
}
\value{
a \link[tibble:tibble-package]{tibble} containing the following columns:
\item{method}{The critical power method for that estimation.}
\item{data}{The data you provided.}
\item{model}{A \code{nls} or \code{lm} object. The model used in the critical power fitting.}
\item{CP or CS}{Critical power, in watts, or critical speed, in meters per second.}
\item{CP SEE or CS SEE}{Standard error of the estimation of critical power, in watts, or critical speed, in meters per second.}
\item{W' or D'}{Anaerobic work capacity, in joules, or anaerobic distance capacity, in meters.}
\item{W' or D' SEE}{Standard error of the estimation of anaerobic work capacity, in joules, or anaerobic distance capacity, in meters.}
\item{Pmax or Smax}{Maximal instantaneous power/speed, in watts/meters per second (only available in CP/CS 3-hyp).}
\item{Pmax SEE or Smax SEE}{Standard error of the estimation of maximal instantaneous power/speed, in watts/meters per second (only available in CP/CS 3-hyp).}
\item{R2}{R-squared.}
\item{RMSE}{Root mean square error, the units vary depending on the method, as they represent the units of the response (i.e., y-axis): CP3-hyp in seconds, CP2-hyp in seconds, CPlinear in joules, and CP1/time in watts.}
\item{plot}{The critical power plot.}
}
\description{
These functions fit the following critical power/speed models: \cr
\itemize{
\item 3-parameter hyperbolic model (Morton, 1996). \cr
\item 2-parameter hyperbolic model (Hill, 1993). \cr
\item linear model (Moritani, 1981). \cr
\item 1/time linear model (Whipp et al. 1982). \cr
For more details see \code{?critical_power} or \code{?critical_speed}.
}
}
\keyword{internal}
