% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cp.R
\name{plot_cp}
\alias{plot_cp}
\title{Plot CP model}
\usage{
plot_cp(
  .data,
  power_output_column,
  time_to_exhaustion_column,
  method = c("3-hyp", "2-hyp", "linear", "1/time"),
  model
)
}
\arguments{
\item{.data}{A data frame containing a power output column and their respective time-to-exhaustion.}

\item{power_output_column}{The name of the power output column. This value has to be in watts. Default to \code{"PO"}.}

\item{time_to_exhaustion_column}{The name of the time-to-exhaustion column. This value has to be in seconds. Default to \code{"TTE"}.}

\item{method}{The method for estimating critical power. It can be one of \code{c("3-hyp", "2-hyp", "linear", "1/time")}.}

\item{model}{The \code{nls} or \code{lm} object derived in one of the \code{method_} functions.}
}
\value{
A \code{ggplot2} object
}
\description{
Plot CP model
}
