% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-cp.R
\name{summary_cp}
\alias{summary_cp}
\title{CP modeling summary}
\usage{
summary_cp(
  .data,
  time_to_exhaustion_column,
  model,
  method,
  critical_speed = FALSE
)
}
\arguments{
\item{.data}{A data frame containing a power output column and their respective time-to-exhaustion.}

\item{time_to_exhaustion_column}{The name of the time-to-exhaustion column. This value has to be in seconds. Default to \code{"TTE"}.}

\item{model}{The \code{nls} or \code{lm} object derived in one of the \code{method_} functions.}

\item{method}{The method for estimating critical power. It can be one of \code{c("3-hyp", "2-hyp", "linear", "1/time")}.}

\item{critical_speed}{A boolean to specify whether the estimation of critical speed is being performed, instead of critical power. Default to \code{FALSE}.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Retrieves a summary of the critical power model in a tidy way. Mainly for internal use.
}
